pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "podcast-fetcher"
version = "0.1.0"
description = "A tool to fetch and store podcast episodes from RSS feeds"
readme = "README.md"
requires-python = ">=3.11.9"
authors = [
    { name = "Your Name", email = "your.email@example.com" },
]
license = "MIT"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

dependencies = [
    "beautifulsoup4>=4.13.4",
    "feedparser>=6.0.11",
    "python-dotenv>=1.1.1",
    "sqlmodel>=0.0.24",
    "requests>=2.31.0",
    "joblib>=1.5.1",
    "psycopg2-binary>=2.9.10",
    "llama-index>=0.13.2",
    "llama-index-llms-bedrock-converse>=0.8.2",
    "boto3>=1.39.11",
    "click>=8.2.1",
    "telethon>=1.40.0",
    "python-telegram-bot>=22.5",
    "loguru>=0.7.1",
    "python-dateutil>=2.8.2",
    "uvicorn>=0.37.0",
    "httpx>=0.28.1",
    "fastapi>=0.118.0",
    "strands-agents>=1.12.0",
    "strands-agents-tools>=0.2.11",
    "aiogram>=3.22.0",
    "pydantic-settings>=2.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[project.scripts]
podcast-fetcher = "podcast_fetcher.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v"

[tool.setuptools.packages.find]
include = ["podcast_fetcher*"]
exclude = ["ui*", "bot*", "data*", "aws_lambda_function*"]

[dependency-groups]
dev = [
    "isort>=6.0.1",
    "modal>=1.1.4",
]
